+++ packages/telescope-lib/lib/icons.js
@@ -9,15 +9,15 @@ Telescope.utils.getIcon = function (iconName, iconClass) {
   var icons = Telescope.utils.icons;
   var iconCode = !!icons[iconName] ? icons[iconName] : iconName;
   var iconClass = (typeof iconClass === 'string') ? ' '+iconClass : '';
-  return '<i class="icon fa fa-' + iconCode + ' icon-' + iconName + iconClass+ '" aria-hidden="true"></i>';
+  return '<i class="icon fa fa-fw fa-' + iconCode + ' icon-' + iconName + iconClass+ '" aria-hidden="true"></i>';
 };
 
 /**
  * A directory of icon keys and icon codes
  */
 Telescope.utils.icons = {
-  expand: "plus",
-  collapse: "minus",
+  expand: "angle-down",
+  collapse: "angle-up",
   close: "times",
   upvote: "chevron-up",
   voted: "check",