+++ packages/telescope-notifications/lib/client/templates/notifications_menu.js
@@ -39,13 +39,13 @@ Template.notifications_menu.helpers({
     }
     return menuItems;
   },
-  menuMode: function () {
+  menuClass: function () {
     if (!!this.mobile) {
-      return 'list';
+      return 'menu-collapsible';
     } else if (Settings.get('navLayout', 'top-nav') === 'top-nav') {
-      return 'dropdown';
+      return 'menu-dropdown';
     } else {
-      return 'accordion';
+      return 'menu-collapsible';
     }
   }
 });