+++ packages/telescope-tags/lib/client/templates/categories_menu.js
@@ -22,13 +22,14 @@ Meteor.startup(function () {
       });
       return defaultItem.concat(menuItems);
     },
-    menuMode: function () {
-      if (!!this.mobile) {
-        return 'list';
+    menuClass: function () {
+      // go back up 4 levels to get the zone that's including the menu
+      if (Template.parentData(4).zone === "mobileNav") {
+        return 'menu-collapsible';
       } else if (Settings.get('navLayout', 'top-nav') === 'top-nav') {
-        return 'dropdown';
+        return 'menu-dropdown';
       } else {
-        return 'accordion';
+        return 'menu-collapsible menu-always-expanded';
       }
     }
   });