+++ packages/telescope-tags/lib/callbacks.js
@@ -21,7 +21,8 @@ Telescope.callbacks.add("postClass", addCategoryClass);
 
 // when a category is added to a post, increment counter
 function updateCategoryCountOnSubmit (post) {
-  Categories.update({_id: {$in: post.categories}}, {$inc: {"postsCount": 1}}, {multi: true});
+  if (!_.isEmpty(post.categories))
+    Categories.update({_id: {$in: post.categories}}, {$inc: {"postsCount": 1}}, {multi: true});
 }
 Telescope.callbacks.add("postSubmitAsync", updateCategoryCountOnSubmit);
 
@@ -30,8 +31,11 @@ function updateCategoryCountOnEdit (newPost, oldPost) {
   var categoriesAdded = _.difference(newPost.categories, oldPost.categories);
   var categoriesRemoved = _.difference(oldPost.categories, newPost.categories);
 
-  Categories.update({_id: {$in: categoriesAdded}}, {$inc: {"postsCount": 1}}, {multi: true});
-  Categories.update({_id: {$in: categoriesRemoved}}, {$inc: {"postsCount": -1}}, {multi: true});
+  if (!_.isEmpty(categoriesAdded))
+    Categories.update({_id: {$in: categoriesAdded}}, {$inc: {"postsCount": 1}}, {multi: true});
+  
+  if (!_.isEmpty(categoriesRemoved))
+    Categories.update({_id: {$in: categoriesRemoved}}, {$inc: {"postsCount": -1}}, {multi: true});
 }
 Telescope.callbacks.add("postEditAsync", updateCategoryCountOnEdit);